Develop

+++ src/main/java/ISPP/G5/INDVELOPERS/controllers/GameController.java
@@ -19,8 +19,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import ISPP.G5.INDVELOPERS.models.Developer;
 import ISPP.G5.INDVELOPERS.models.Game;
-import ISPP.G5.INDVELOPERS.repositories.GameRepository;
 import ISPP.G5.INDVELOPERS.services.DeveloperService;
 import ISPP.G5.INDVELOPERS.services.GameService;
 
@@ -32,9 +32,6 @@ public class GameController {
 
 	@Autowired
 	private GameService gameService;
-	
-	@Autowired
-	private GameRepository gameRepository;
 	
 	@Autowired
 	private DeveloperService developerService;
@@ -56,11 +53,19 @@ public class GameController {
 	
 	@PostMapping("/add")
 	public ResponseEntity<String> addGame(@RequestBody Game game) throws NotFoundException {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = this.developerService.findByUsername(userDetails.getUsername());
+		boolean isPremium = false;
+		if(developer.getIsPremium() != null) {
+			isPremium = developer.getIsPremium();
+		}
 		try {
-			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-			UserDetails userDetails = (UserDetails) authentication.getPrincipal();
-			game.setCreator(this.developerService.findByUsername(userDetails.getUsername()));
-			return ResponseEntity.status(HttpStatus.CREATED).body(this.gameService.addGame(game));
+			if(isPremium == false && game.getPrice() != 0.0) 
+				throw new IllegalArgumentException("Only premium developers can sell non-free games");
+			if(isPremium == false && (this.gameService.findByMyGames(developer.getId()).size() + 1 == 6))
+				throw new IllegalArgumentException("Non premium developers only can have a maximun of five games published");
+			return ResponseEntity.status(HttpStatus.CREATED).body(this.gameService.addGame(game, developer));
 		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
@@ -68,8 +73,13 @@ public class GameController {
 	
 	@PutMapping("/edit/{id}")
 	public ResponseEntity<String> updateGame(@PathVariable String id, @RequestBody Game game) throws NotFoundException{
-		Game gameData = this.gameService.findById(id);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = developerService.findByUsername(userDetails.getUsername());
+		Game gameData = gameService.findById(id);
 		try {
+			if (!game.getCreator().getId().equals(developer.getId())) 
+				throw new IllegalArgumentException("Only the creator of the game can edit it");
 			gameData.setTitle(game.getTitle());
 			gameData.setDescription(game.getDescription());
 			gameData.setRequirements(game.getRequirements());
@@ -84,28 +94,36 @@ public class GameController {
 	
 	@DeleteMapping("/delete/{id}")
 	public ResponseEntity<HttpStatus> deleteGameById(@PathVariable("id") String id) throws NotFoundException{
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = this.developerService.findByUsername(userDetails.getUsername());
+		Game game = this.gameService.findById(id);
 		try {
+			if (!game.getCreator().getId().equals(developer.getId())) { 
+				throw new IllegalArgumentException("Only the creator of the game can remove it");
+			}
 			this.gameService.deleteGame(id);
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 		} catch(IllegalArgumentException e) {
-			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 	}
 	
 	@GetMapping("/findByTitle/{title}")
-	public ResponseEntity<Game> getGameByTitle(@PathVariable String title) {
+	public ResponseEntity<List<Game>> getGameByTitle(@PathVariable String title) {
 		try {
 			return ResponseEntity.ok(this.gameService.findByTitle(title));
-		} catch(IllegalArgumentException | NotFoundException e) {
+		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
 	
 	@GetMapping("/findByDeveloper/{developerUsername}")
-	public ResponseEntity<List<Game>> getGameByDeveloper(@PathVariable String developerUsername) {
+	public ResponseEntity<List<Game>> getGameByDeveloper(@PathVariable String developerUsername) throws NotFoundException {
 		try {
-			return ResponseEntity.ok(this.gameService.findByDeveloper(developerUsername));
-		} catch(IllegalArgumentException | NotFoundException e) {
+			Developer developer = this.developerService.findByUsername(developerUsername);
+			return ResponseEntity.ok(this.gameService.findByDeveloper(developer.getId()));
+		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
@@ -113,8 +131,20 @@ public class GameController {
 	@GetMapping("/findMyGames")
 	public ResponseEntity<List<Game>> getGameByMyGames() {
 		try {
-			return ResponseEntity.ok(this.gameService.findByMyGames());
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+			Developer developer = this.developerService.findByUsername(userDetails.getUsername());
+			return ResponseEntity.ok(this.gameService.findByMyGames(developer.getId()));
 		} catch(IllegalArgumentException | NotFoundException e) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+		}
+	}
+	
+	@GetMapping("/{id}")
+	public ResponseEntity<Game> getGameById(@PathVariable String id) {
+		try {
+			return ResponseEntity.ok(this.gameService.findById(id));
+		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
